move upload workflows modal to a separate component
